Improve customer notification subscription form
Previously, product out-of-stock subscription form was displayed only
when it was not possible to order the product. When product was actually
out-of-stock, but back orders were allowed, the form was not displayed.

This commit allows merchants to decide between this functionality, or to
always display subscription form when product is not in stock (even if
it can be back-ordered)