Do not try to detect email directory
The code that checks existence of email template can lead to emails not
being sent. Ror example, when txt variant of email template does not
exists, no email is sent.

Module should always attempt to sent email -- it's up to core to find
email template, and log error if it's missing.